migrating prepareFlexFunction over to function-helper
