+++ serverless-functions/src/functions/features/chat-to-video-escalation/agent-get-token.js
@@ -1,7 +1,7 @@
 const AccessToken = require('twilio').jwt.AccessToken;
 
 const { SyncGrant, VideoGrant } = AccessToken;
-const { prepareFlexFunction } = require(Runtime.getFunctions()['common/helpers/prepare-function'].path);
+const { prepareFlexFunction } = require(Runtime.getFunctions()['common/helpers/function-helper'].path);
 const SyncOperations = require(Runtime.getFunctions()['common/twilio-wrappers/sync'].path);
 
 const requiredParameters = [{ key: 'DocumentSid', purpose: 'used for sync document' }];
